backend/app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.2
djangorestframework==3.14.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
multidict==6.0.5
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.1.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.42.0
yarl==1.9.4
